+++ modules/prelude-programming.el
@@ -65,7 +65,8 @@
              prelude-flyspell)
     (flyspell-prog-mode))
   (when prelude-guru
-    (guru-mode +1))
+    (guru-mode +1)
+    (diminish 'guru-mode))
   (smartparens-mode +1)
   (prelude-enable-whitespace)
   (prelude-local-comment-auto-fill))