+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 * Added `transient/` to `.gitignore`.
 * Fallback to `sample/prelude-modules.el` in the absence of a `prelude-modules.el` in one's personal folder.
 * [Ruby] Don't auto-insert coding comments.
+* Hide (via `diminish`) `editorconfig-mode`, `super-save`, `company`, `abbrev` and `ivy` in the modeline.
 
 ### Bugs fixed